+++ src/App/PropertyStandard.cpp
@@ -478,7 +478,7 @@ void PropertyEnumeration::setPyObject(PyObject *value)
             if(seq.size() == 2) {
                 Py::Object v(seq[0].ptr());
                 if(!v.isString() && v.isSequence()) {
-                    idx = Py::Int(seq[1].ptr());
+                    idx = Py::Long(seq[1].ptr());
                     seq = v;
                 }
             }
@@ -579,14 +579,14 @@ bool PropertyEnumeration::getPyPathValue(const ObjectIdentifier &path, Py::Objec
         else {
             Py::Tuple tuple(2);
             tuple.setItem(0, res);
-            tuple.setItem(1, Py::Int(getValue()));
+            tuple.setItem(1, Py::Long(getValue()));
             r = tuple;
         }
     } else if (p == ".String") {
         auto v = getValueAsString();
         r = Py::String(v?v:"");
     } else
-        r = Py::Int(getValue());
+        r = Py::Long(getValue());
     return true;
 }
